firewall: reduce MTU for IPv6 to 1460
There is no logic in this change. A MTU of 1480 is the normal value. 6rd
is adding an IPv4 header, so 20 bytes. However, some destinations seem
to have become unreachable randomly with such a MTU. Lowering the MTU to
1460 fix the problem. Also, TCP-MSS to 1440 (1480 - 40) does NOT fix the
problem. So, there is really something which seems to not allow packets
too big and not compatible with PMTU.